feature: sign up refactor domain package
