+++ app/src/main/java/com/hacka/presenteperfeito/feature/signup/domain/repository/SignUpRepository.kt
@@ -1,4 +1,4 @@
-package com.hacka.presenteperfeito.feature.signup.data.repository
+package com.hacka.presenteperfeito.feature.signup.domain.repository
 
 import com.hacka.presenteperfeito.feature.signup.data.model.SignUpReqDTO
 import com.hacka.presenteperfeito.feature.signup.data.model.SignUpResDTO
