+++ app/src/main/java/com/hacka/presenteperfeito/feature/signup/domain/usecase/SignUpUseCase.kt
@@ -1,6 +1,6 @@
-package com.hacka.presenteperfeito.feature.signup.data.usecase
+package com.hacka.presenteperfeito.feature.signup.domain.usecase
 
-import com.hacka.presenteperfeito.feature.signup.data.repository.SignUpRepository
+import com.hacka.presenteperfeito.feature.signup.domain.repository.SignUpRepository
 import com.hacka.presenteperfeito.feature.signup.presentation.model.CreateUser
 import kotlinx.coroutines.flow.flow
 import org.koin.core.annotation.Factory
